requirements.txt
GitPython>=3.1.0
PyGithub>=2.1.0
Flask>=3.0.0
flask-cors>=4.0.0
pyngrok>=7.0.0
Pillow>=10.0.0
qrcode[pil]>=7.4
requests>=2.31.0
python-telegram-bot>=20.0
